refactor : change the structure of folders; style : rename classe and their interfaces
