+++ bot/src/main/java/edu/java/bot/commands/TrackCommand.java
@@ -2,7 +2,7 @@ package edu.java.bot.commands;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.validators.ValidatorRepo;
+import edu.java.bot.services.validators.ValidatorRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
