+++ bot/src/main/java/edu/java/bot/services/listeners/IUpdatesListenerBot.java
@@ -1,4 +1,4 @@
-package edu.java.bot;
+package edu.java.bot.services.listeners;
 
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
@@ -7,7 +7,7 @@ import com.pengrad.telegrambot.response.BaseResponse;
 import jakarta.annotation.PostConstruct;
 import java.util.List;
 
-public interface Bot extends AutoCloseable, UpdatesListener {
+public interface IUpdatesListenerBot extends AutoCloseable, UpdatesListener {
     <T extends BaseRequest<T, R>, R extends BaseResponse> void execute(BaseRequest<T, R> request);
 
     @Override
