+++ bot/src/main/java/edu/java/bot/services/processors/UserMessageProcessorImpl.java
@@ -1,4 +1,4 @@
-package edu.java.bot.processor;
+package edu.java.bot.services.processors;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
