+++ bot/src/test/java/edu/java/bot/CommandsTest.java
@@ -7,10 +7,10 @@ import edu.java.bot.commands.Command;
 import edu.java.bot.commands.HelpCommand;
 import edu.java.bot.commands.StartCommand;
 import edu.java.bot.commands.TrackCommand;
-import edu.java.bot.validators.GitHubValidator;
-import edu.java.bot.validators.LinkValidator;
-import edu.java.bot.validators.StackOverFlowValidator;
-import edu.java.bot.validators.ValidatorRepo;
+import edu.java.bot.services.validators.GitHubValidator;
+import edu.java.bot.services.validators.LinkValidator;
+import edu.java.bot.services.validators.StackOverFlowValidator;
+import edu.java.bot.services.validators.ValidatorRepo;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
