+++ scrapper/src/main/java/edu/java/services/clients/BotClient.java
@@ -1,7 +1,30 @@
 package edu.java.services.clients;
 
 import edu.java.models.dto.request.LinkUpdate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
 
-public interface BotClient {
-    String update(LinkUpdate update);
+@Service
+public class BotClient implements IBotClient {
+    private final static String ENDPOINT = "/updates";
+    private final WebClient webClient;
+
+    @Autowired
+    public BotClient(@Qualifier("botClient") WebClient webClient) {
+        this.webClient = webClient;
+    }
+
+    @Override
+    public String update(LinkUpdate linkUpdate) {
+        return webClient
+            .post()
+            .uri(ENDPOINT)
+            .body(BodyInserters.fromValue(linkUpdate))
+            .retrieve()
+            .bodyToMono(String.class)
+            .block();
+    }
 }
