+++ scrapper/src/main/java/edu/java/services/clients/GitHubClient.java
@@ -1,7 +1,26 @@
 package edu.java.services.clients;
 
 import edu.java.models.GitHubRepoLastUpdate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.client.WebClient;
 
-public interface GitHubClient {
-    GitHubRepoLastUpdate fetchRepo(String user, String repository);
+@Service
+public class GitHubClient implements IGitHubClient {
+    private final WebClient webClient;
+
+    @Autowired
+    public GitHubClient(@Qualifier("gitHubClient") WebClient webClient) {
+        this.webClient = webClient;
+    }
+
+    @Override
+    public GitHubRepoLastUpdate fetchRepo(String user, String repository) {
+        return webClient.get()
+            .uri("/repos/{user}/{repository}", user, repository)
+            .retrieve()
+            .bodyToMono(GitHubRepoLastUpdate.class)
+            .block();
+    }
 }
