+++ scrapper/src/main/java/edu/java/services/clients/IBotClient.java
@@ -1,0 +1,7 @@
+package edu.java.services.clients;
+
+import edu.java.models.dto.request.LinkUpdate;
+
+public interface IBotClient {
+    String update(LinkUpdate update);
+}
