+++ scrapper/src/main/java/edu/java/services/clients/IGitHubClient.java
@@ -1,0 +1,7 @@
+package edu.java.services.clients;
+
+import edu.java.models.GitHubRepoLastUpdate;
+
+public interface IGitHubClient {
+    GitHubRepoLastUpdate fetchRepo(String user, String repository);
+}
