+++ scrapper/src/main/java/edu/java/services/clients/IStackOverFlowClient.java
@@ -1,0 +1,7 @@
+package edu.java.services.clients;
+
+import edu.java.models.StackOverFlowLastUpdate;
+
+public interface IStackOverFlowClient {
+    StackOverFlowLastUpdate fetchQuestion(String questionId);
+}
