+++ scrapper/src/main/java/edu/java/services/scheduler/LinkUpdaterScheduler.java
@@ -1,11 +1,11 @@
-package edu.java.scheduler;
+package edu.java.services.scheduler;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-@Component
+@Service
 @EnableScheduling
 @Log4j2
 public class LinkUpdaterScheduler {
