+++ scrapper/src/test/java/edu/java/scrapper/IntegrationTest.java
@@ -1,12 +1,17 @@
 package edu.java.scrapper;
 
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.nio.file.Path;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import liquibase.Contexts;
 import liquibase.LabelExpression;
 import liquibase.Liquibase;
 import liquibase.database.Database;
 import liquibase.database.DatabaseFactory;
 import liquibase.database.jvm.JdbcConnection;
-import liquibase.exception.DatabaseException;
 import liquibase.exception.LiquibaseException;
 import liquibase.resource.DirectoryResourceAccessor;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -14,12 +19,6 @@ import org.springframework.test.context.DynamicPropertySource;
 import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.nio.file.Path;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
 
 @Testcontainers
 public abstract class IntegrationTest {
@@ -50,9 +49,11 @@ public abstract class IntegrationTest {
                 .getInstance()
                 .findCorrectDatabaseImplementation(new JdbcConnection(connection));
 
-            Liquibase liquibase = new liquibase.Liquibase("master.xml",
+            Liquibase liquibase = new liquibase.Liquibase(
+                "master.xml",
                 new DirectoryResourceAccessor(changelogPath),
-                database);
+                database
+            );
             liquibase.update(new Contexts(), new LabelExpression());
         } catch (SQLException | FileNotFoundException | LiquibaseException e) {
             throw new RuntimeException(e);
